+++ RenderWindow.hpp
@@ -22,7 +22,7 @@ public:
 	void setupImGui();
 	void renderTestImGuiText();
 	void drawImgui();
-	void renderGrid(int width, int height, int cameraX, int cameraY, int zoomLevel, SDL_Texture* texture);
+	void renderGrid(int width, int height, float cameraX, float cameraY, int zoomLevel, SDL_Texture* texture);
 private:
 	SDL_Window* window;
 	SDL_Renderer* renderer;
